Extract fixtures.py for networking
