+++ tests/integration/networking/test_networking.py
@@ -8,77 +8,20 @@ from _pytest.monkeypatch import MonkeyPatch
 from tests.integration.helpers import (
     BASE_CMDS,
     assert_headers_in_lines,
-    delete_target_id,
     exec_test_command,
 )
-from tests.integration.linodes.helpers import (
-    create_linode,
-    create_linode_and_wait,
-    DEFAULT_REGION
+from tests.integration.linodes.helpers import DEFAULT_REGION
+from tests.integration.networking.fixtures import (
+    create_reserved_ip,
+    get_command_heads_and_vals,
+    test_linode_id,
+    test_linode_id_shared_ipv4
 )
-from tests.integration.sharegroups.fixtures import get_region  # noqa: F401
 
 
 RESERVED_IP_HEADERS = [
     "address", "type", "public", "rdns", "linode_id", "reserved", "tags"
 ]
-
-
-@pytest.fixture
-def create_reserved_ip(request):
-    tags = getattr(request, "param", None)
-    command = BASE_CMDS["networking"] + [
-        "reserved-ip-add",
-        "--region",
-        DEFAULT_REGION,
-        "--text",
-        "--delimiter",
-        ","
-    ]
-
-    if tags:
-        command += ["--tags", tags]
-
-    headers, values = get_command_heads_and_vals(command)
-    yield headers, values
-
-    delete_target_id("networking", values[0], "reserved-ip-delete")
-
-
-@pytest.fixture(scope="package")
-def test_linode_id(linode_cloud_firewall):
-    linode_id = create_linode_and_wait(firewall_id=linode_cloud_firewall)
-
-    yield linode_id
-
-    delete_target_id(target="linodes", id=linode_id)
-
-
-@pytest.fixture(scope="package")
-def test_linode_id_shared_ipv4(linode_cloud_firewall):
-    target_region = "us-mia"
-
-    linode_ids = (
-        create_linode(
-            test_region=target_region, firewall_id=linode_cloud_firewall
-        ),
-        create_linode(
-            test_region=target_region, firewall_id=linode_cloud_firewall
-        ),
-    )
-
-    yield linode_ids
-
-    for id in linode_ids:
-        delete_target_id(target="linodes", id=id)
-
-
-def get_command_heads_and_vals(command):
-    result = exec_test_command(command).splitlines()
-    headers = [item for item in result[0].split(",")]
-    values = [item for item in result[1].split(",")]
-
-    return headers, values
 
 
 def has_shared_ip(linode_id: int, ip: str) -> bool:
